@@ -1,24 +1,20 @@
 {
-  "name": "jenkins-2",
+  "name": "jenkins-prototype",
   "version": "0.1.0",
-  "engines": {
-    "node": "16.17.0",
-    "npm": "8.5.5"
-  },
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@tippyjs/react": "^4.2.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4",
     "react-ionicons": "^4.2.0",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
     "recharts": "^2.1.14",
     "sass": "^1.42.1",
-    "web-vitals": "^1.0.1"
+    "@tippyjs/react": "^4.2.5",
+    "react-router-dom": "^6.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
